feat: :sparkles: 333. Deleting Files

+++ controllers/admin.js
@@ -1,5 +1,5 @@
 const { validationResult } = require('express-validator');
-
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 
 /**
@@ -176,6 +176,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (updatedImage){
+                //fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = updatedImage.path;
             }
             return product.save()
@@ -194,14 +195,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-        console.log('Producto eliminado exitosamente!');
+    Product.findById(prodId)
+      .then(product => {
+        if (!product) {
+          return next(new Error('Product not found.'));
+        }
+        //fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+      })
+      .then(() => {
+        console.log('DESTROYED PRODUCT');
         res.redirect('/admin/products');
-    })
-    .catch(err =>{
+      })
+      .catch(err => {
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
-    });
-};
+      });
+  };
